Cut unneeded infrastructure
